x86 boot: fix initial pages region size
Resetting the size of the initial pages region to
INITIAL_PAGES_SIZE after doing the initial mapping is no longer
needed (and must be avoided) after the changes brought by #1507,
otherwise some pages of physical memory that have been used for the
initial mapping may be re-used for subsequent mappings, thereby
corrupting the initial mapping.
This commit fixes the above issue (which has been observed when
running the tlbshootdown runtime test) for Firecracker and UEFI
boot, by removing the code that resets the inital pages region,
analogously to what commit 9b6268a24c2918e442f7dbc9345371c60d9edd67
did for stage2. The static variable pointing to the initial pages
region in stage2 is being removed because it was only needed when
the region size had to be reset.