x86 boot: add support for Qemu direct kernel boot
This change adds a new entry point to the kernel, which implements
the x86 HVM direct boot protocol and is used by Qemu when the
`-kernel` command line option is supplied. The location of this
entry point (where the kernel runs in 32-bit mode without paging)
is specified in the kernel binary file by means of a special ELF
note.
The KERNEL_BOOT Makefile variable is used to specify whether direct
kernel boot is used when running a unikernel image, as in
`make run KERNEL_BOOT=1`. Running an image with direct kernel boot
results in faster boot times.